Introduce UnityMenuModel and UnityQmlMenuModel
UnityMenuModel serves the same purpose as QMenuModel, but it is based on
GtkMenuTracker and has a different API.

GtkMenuTracker is maintained in gtk, but meant to be copy-and-pasted into other
consumers of menu models. It does not introduce new dependencies. It does give
us access to all the features that GMenuModel has.